+++ src/vendors/onedrive/onedrive_client.c
@@ -269,7 +269,7 @@ static int oauth_writeback(const cJSON *json, void *user_data)
         return HIVE_SYS_ERROR(errno);
     }
 
-    json_str_len = strlen(json_str);
+    json_str_len = (int)strlen(json_str);
     bytes = (int)write(fd, json_str, json_str_len + 1);
     free(json_str);
     close(fd);
